Use the preexisting `EraPrice` struct
